adapt function param

+++ src/cheat.c
@@ -1229,7 +1229,7 @@ static void		DoCheatEntry(CheatEntry * entry);
 static void		UpdateAllCheatInfo(void);
 static void		UpdateCheatInfo(CheatEntry * entry, UINT8 isLoadTime);
 
-static int		IsAddressInRange(CheatAction * action, UINT32 length);
+static int		IsAddressInRange(CheatAction * action, size_t length);
 
 static void		BuildCPUInfoList(void);
